fix: refactoring Comment component
